requirements.txt
streamlit>=1.30.0
PyGithub>=2.1.0
Pillow>=10.0.0
